chore: merge unit tests for NegotiationController into integration tests
